requirements.txt
requests==2.30.0
python-environ==0.4.54